+++ app/charts/map/map-custom-layers-legend.tsx
@@ -2,7 +2,7 @@ import { Box, Typography } from "@mui/material";
 import uniq from "lodash/uniq";
 import NextImage from "next/image";
 
-import { useWMTSorWMSLayers } from "@/charts/map/wms-endpoint-utils";
+import { useWMTSorWMSLayers } from "@/charts/map/wms-wmts-endpoint-utils";
 import { DEFAULT_WMS_URL, ParsedWMSLayer } from "@/charts/map/wms-utils";
 import {
   DEFAULT_WMTS_URL,
